w'day/requirements.txt
pyodbc==5.1.0
pandas==2.2.0
python-dotenv==1.0.1
